+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export * from "./teams";
 export * from "./gmail";
 export * from "./finance";
 export * from "./notion";
+export * from "./pmtool";